Fix markdown headers in resume
There was a quirk rendering markdown in which an h4 element had a smaller font
than an h5 element. This did not match the local markdown preview. In an effort
to correct this I have bumped the h4 elements up h3. It's also possible that
my markdown formatting is incorrect and styles are leaking into places they
shouldn't be.