Fix a broken favicon path
The favicon was rendering as expected on index.html, but not on any of the
other markdown pages. This was happening because of a relative path which was
only valid from the root directoy. When the site gets rendered by Jekyll, it
appears that the structure changes slightly, e.g. resume.md ->
nathanson.dev/resume/, so this path didn't work once we navigated to the resume
page.

This was fixed by replacing that relative path with an absolute path. The
disadvantage of this is technical debt if I ever need to change the website
name or restructure the assets for some reasons.