Fix and prevent future mixed HTTP/HTTPS content.
While working on the music section, I used a service to auto-generate markdown
images and links for the preview of a YouTube video. This worked exactly as
intended. However, the service used "http:" rather than "https:". This one
character difference is easy to miss but important in terms of privacy (the
latter uses encryption and prevents visitors to my site from being snooped on
easily). So when I published the page, Chrome/FireFox/Safari reported that this
was an "insecure" page due to mixed use of http/https throughout the content.

To resolve this, I replaced any "http" links in the markdown side of the site
as well as (most of) the docs.

To prevent this from happening again, I broadened the scope of the Dead Linke
Checker to also grep for "http:" in top-level markdown files.